+++ lib/policyBuilder.ts
@@ -1,6 +1,6 @@
 import { createMongoAbility, MongoAbility, RawRuleOf, InferSubjects } from "@casl/ability";
 import { FormRecord } from "@lib/types";
-import { User, Privelege as PrismaPrivelege } from "@prisma/client";
+import { User, Privilege as PrismaPrivilege } from "@prisma/client";
 /*
 This file is referenced by the useAccessControl hook so no server-side
 only dependencies can be referenced in this file.
@@ -14,13 +14,13 @@ interface CASL_User extends User {
   kind: "User";
 }
 
-interface CASL_Privelege extends Privelege {
-  kind: "Privelege";
+interface CASL_Privilege extends Privilege {
+  kind: "Privilege";
 }
 
 export type Action = "manage" | "create" | "view" | "update" | "delete";
 
-export type Subject = InferSubjects<CASL_FormRecord | CASL_User | CASL_Privelege>;
+export type Subject = InferSubjects<CASL_FormRecord | CASL_User | CASL_Privilege>;
 
 export type Abilities = [Action, Subject];
 export type AppAbility = MongoAbility<Abilities>;
@@ -33,7 +33,7 @@ export type Permission = {
   subject: Extract<Subject, string> | Extract<Subject, string>[];
   condition?: string;
 };
-export interface Privelege extends PrismaPrivelege {
+export interface Privilege extends PrismaPrivilege {
   permissions: Permission[];
   [key: string]: string | null | Permission[];
 }