+++ pages/api/privelages.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { middleware, cors, sessionExists } from "@lib/middleware";
-import { getAllPriveleges } from "@lib/priveleges";
+import { getAllPrivileges } from "@lib/privileges";
 
 import { AdminLogAction } from "@lib/adminLogs";
 import { Session } from "next-auth";
@@ -8,34 +8,34 @@ import { MiddlewareProps } from "@lib/types";
 import { logMessage } from "@lib/logger";
 import { createAbility, Ability } from "@lib/policyBuilder";
 import {
-  updatePrivelege as prismaUpdatePrivelege,
-  createPrivelege as prismaCreatePrivelege,
-} from "@lib/priveleges";
+  updatePrivilege as prismaUpdatePrivilege,
+  createPrivilege as prismaCreatePrivilege,
+} from "@lib/privileges";
 const allowedMethods = ["GET", "PUT", "POST"];
 
-const getPrivelegeList = async (res: NextApiResponse, ability: Ability) => {
-  const priveleges = await getAllPriveleges(ability);
-  if (priveleges.length === 0) {
+const getPrivilegeList = async (res: NextApiResponse, ability: Ability) => {
+  const privileges = await getAllPrivileges(ability);
+  if (privileges.length === 0) {
     res.status(500).json({ error: "Could not process request" });
   } else {
-    res.status(200).json([...priveleges]);
+    res.status(200).json([...privileges]);
   }
 };
 
-const createPrivelege = async (
+const createPrivilege = async (
   req: NextApiRequest,
   res: NextApiResponse,
   ability: Ability,
   session?: Session
 ) => {
-  const { privelege } = req.body;
-  if (typeof privelege === "undefined" || privelege.id) {
+  const { privilege } = req.body;
+  if (typeof privilege === "undefined" || privilege.id) {
     return res.status(400).json({ error: "Malformed Request" });
   }
   // Need to parse serialized permissions from form response
-  privelege.permissions = JSON.parse(privelege.permissions);
+  privilege.permissions = JSON.parse(privilege.permissions);
 
-  const result = await prismaCreatePrivelege(ability, privelege);
+  const result = await prismaCreatePrivilege(ability, privilege);
   logMessage.info(AdminLogAction.Update);
   if (result) {
     if (session && session.user.id) {
@@ -47,24 +47,24 @@ const createPrivelege = async (
     }
     res.status(200).send("Success");
   } else {
-    res.status(404).json({ error: "Privelege not found" });
+    res.status(404).json({ error: "Privilege not found" });
   }
 };
 
-const updatePrivelege = async (
+const updatePrivilege = async (
   req: NextApiRequest,
   res: NextApiResponse,
   ability: Ability,
   session?: Session
 ) => {
-  const { privelege } = req.body;
-  if (typeof privelege === "undefined" || !privelege.id) {
+  const { privilege } = req.body;
+  if (typeof privilege === "undefined" || !privilege.id) {
     return res.status(400).json({ error: "Malformed Request" });
   }
   // Need to parse serialized permissions from form response
-  privelege.permissions = JSON.parse(privelege.permissions);
+  privilege.permissions = JSON.parse(privilege.permissions);
 
-  const result = await prismaUpdatePrivelege(ability, privelege);
+  const result = await prismaUpdatePrivilege(ability, privilege);
 
   logMessage.info(AdminLogAction.Update);
   if (result) {
@@ -77,7 +77,7 @@ const updatePrivelege = async (
     }
     res.status(200).send("Success");
   } else {
-    res.status(404).json({ error: "Privelege not found" });
+    res.status(404).json({ error: "Privilege not found" });
   }
 };
 
@@ -91,17 +91,17 @@ const handler = async (
       res.status(403);
       return;
     }
-    const ability = createAbility(session.user.priveleges);
+    const ability = createAbility(session.user.privileges);
 
     switch (req.method) {
       case "POST":
-        await createPrivelege(req, res, ability, session);
+        await createPrivilege(req, res, ability, session);
         break;
       case "PUT":
-        await updatePrivelege(req, res, ability, session);
+        await updatePrivilege(req, res, ability, session);
         break;
       case "GET":
-        await getPrivelegeList(res, ability);
+        await getPrivilegeList(res, ability);
         break;
     }
   } catch (error) {